Create Mutli GeneLists and Bedgraphs
This has been rewritten to allow multiple gene lists and bedgraph files to be examined without the need to rerun the code. Users must give the path to the gtf file, bedgraph, and gene lists of interest. While designed for Arabidopsis gene annotation AT#G#####, this can be modified to use with other organisms.