Create meanPlotFunction
This function is inspired by DESeq2's file input. Here, two 2 by x tables are made, one containing gene lists of interest and a description of the file, the other containing bedgraphs and a description of the file. The meanPlot function requires 4 inputs, the gene list table, the bedgraph table, a point upstream of the start codon (start), and a point downstream of the start codon (end). Additionally, users can specify an input file to normalize IP data against over gene regions.